Initial commit - web based command line.
This is a client-side tool for providing command-line interfaces over the web.
Think of a web-enabled mathematica notebook or interpreter shell as the kind of
application this might be good for.

command.js   -- Javascript code implementing the command line UI.
command.html -- html shell for pulling in command.js (and other bits).
jabber.html  -- The Jabberwocky.