Not just the Jabberwocky.
We now actually parse the command line a little and fetch the URI provided
there.  If the fetch fails, the exception is caught a and the result printed in
red (although this might not work over the network: there we will get an HTTP
error code instead of an exception).