Folding Panes.
You can now click on old command bars, and the response pane below it will
disappear, click on it again, and it will reappear.  This was suprisingly fun
and easy to implement.  I think adding other manipulations will be easy to do
as they become desired, so I have taken that item off the TODO list as well.