REFACTOR: requestURI no longer touches the receiver.
Now requestURI returns a new DOM object to hold the data and runCommand inserts
into the the receiver.  This means we don't have to pass requestURI a receiver
arg, thus loosening the depenencies.  In the next step, we will not do
the stupid nesting of OLD_RECEIVER( NEW_RETURNED_OBJECT ).