BUGFIX: Snapshot prompt no longer expands replace empty banners.
Actually, a work around a Webkit bug.

I think there is some deep CSS black magic going on, when I changed the
stylesheet, snaps started breaking when there was no text in the banner.  This
happens on Chromium, but not Firefox.  It happens only to the snaps, not to the
live input.  The difference is related to a CSS property called
"display:inline-block", which might be a Webkit-only thing.

Anyway, my fix is to not have empty banners.  If the text is empty contains no
non-whitespace characters, we shove in an empty <span></span>.  This fixes.
WTF?!

I was reluctant to do this, because it seemed like a hack, especially because
the problem seems to be in the stylesheet.  Until I found how deeply weird, and
probably fuzzy, the stylesheeting is.  Anyway, firefox is fine with the
solution.