REFACTOR: Input objects are now shorter lived.
The a new Input object is now created for every command, although subsequent
Inputs share the same underlying DOM input element.  I am not yet taking full
advantage of this change.

One difficulty which had led to bugs was knowing over what stretches of time
the state of Session is valid.  Now we can say things clearly:

        -> While a user is entering a command, the Session is frozen except it
           points to an Input element which is changing as the user types.

        -> Once the user hits Enter, the Input element is frozen but the
           session element may evolve.

The upshot of this is that Panes should *not* point to session, the should
point to Input, which they know is frozen.